ReCo FE Code Snippets
Challenges
1. Data Synchronization for Overlay Display
Given that this is an overlay on another screen, I needed to be extra careful with the data to ensure it was being displayed correctly in the flatlist. 

2. Handling Various Image Functionalities
Each image within the overlay had various functionalities (e.g., adding to a new collection, opening comments modal, editing caption, etc.). Ensuring the state data was accurate became crucial for ensuring correct behavior.

3. Backend and Frontend sync
I used axios and redux to update data on both the backend and frontend